refactor(reposicao): extrai SugestaoCard/RankingTable/dialogs de NegociacaoParalela
God-component split do AdminReposicaoNegociacaoParalela (1073 -> 702 linhas).
Extração behavior-preserving de JSX presentacional para 3 módulos novos:
- SugestaoCard.tsx (card de sugestão ativa)
- RankingTable.tsx (tabela do ranking completo + loading/empty)
- dialogs.tsx (IgnorarDialog, FecharSemAcordoDialog, ConverterDialog)

Página continua dona de todo estado/queries/handlers; sub-componentes recebem
dados + callbacks via props. Novo type ConvertForm em types.ts. Todos os 5
arquivos (page + 3 módulos + types) ficam em tsconfig.strict.json.

Codex review: behavior-preserving, strict TS passa.
