+++ src/pages/AdminReposicaoNegociacaoParalela.tsx
@@ -9,18 +9,13 @@ import {
   RefreshCw,
   Sparkles,
   ClipboardList,
-  Eye,
-  MoreVertical,
   Search,
   Info,
 } from "lucide-react";
 import { Button } from "@/components/ui/button";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { Card, CardContent } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
 import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Textarea } from "@/components/ui/textarea";
-import { Progress } from "@/components/ui/progress";
 import {
   Select,
   SelectContent,
@@ -29,40 +24,13 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
-import {
   DropdownMenu,
   DropdownMenuContent,
-  DropdownMenuItem,
   DropdownMenuTrigger,
   DropdownMenuCheckboxItem,
   DropdownMenuLabel,
   DropdownMenuSeparator,
 } from "@/components/ui/dropdown-menu";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from "@/components/ui/alert-dialog";
 import {
   Breadcrumb,
   BreadcrumbItem,
@@ -83,17 +51,20 @@ import {
   type OrdenacaoKey,
   type Sugestao,
   type RankingRow,
+  type ConvertForm,
 } from "@/components/reposicao/negociacaoParalela/types";
 import {
-  formatBRL,
-  formatPerc,
   categoriaBadgeClass,
   categoriaLabel,
-  statusBadgeClass,
-  statusLabel,
-  percPromoBadgeClass,
   lastDayOfNextMonth,
 } from "@/components/reposicao/negociacaoParalela/helpers";
+import { SugestaoCard } from "@/components/reposicao/negociacaoParalela/SugestaoCard";
+import { RankingTable } from "@/components/reposicao/negociacaoParalela/RankingTable";
+import {
+  IgnorarDialog,
+  FecharSemAcordoDialog,
+  ConverterDialog,
+} from "@/components/reposicao/negociacaoParalela/dialogs";
 
 export default function AdminReposicaoNegociacaoParalela() {
   const navigate = useNavigate();
@@ -126,13 +97,13 @@ export default function AdminReposicaoNegociacaoParalela() {
   const [fecharSemAcordoTarget, setFecharSemAcordoTarget] = useState<Sugestao | null>(null);
   const [fecharObs, setFecharObs] = useState("");
   const [convertTarget, setConvertTarget] = useState<Sugestao | null>(null);
-  const [convertForm, setConvertForm] = useState({
+  const [convertForm, setConvertForm] = useState<ConvertForm>({
     desconto_perc: 5,
     volume_minimo: 1000,
-    volume_unidade: "reais" as "unidades" | "reais" | "kg" | "litros",
+    volume_unidade: "reais",
     data_fim: lastDayOfNextMonth(),
     responsavel: "",
-    canal: "ligacao" as "email" | "whatsapp" | "ligacao" | "visita_presencial" | "outro",
+    canal: "ligacao",
     observacoes: "",
   });
   const [convertSubmitting, setConvertSubmitting] = useState(false);
@@ -239,8 +210,6 @@ export default function AdminReposicaoNegociacaoParalela() {
     for (const r of ranking) m.set(r.sku_codigo_omie, r);
     return m;
   }, [ranking]);
-
-  // Faltam para próxima faixa não se aplica aqui (era no DES). Removido.
 
   // Actions
   const handleGerarSugestoes = async () => {
@@ -414,125 +383,6 @@ export default function AdminReposicaoNegociacaoParalela() {
     return next;
   };
 
-  // Render helpers
-  const renderSugestaoCard = (s: Sugestao) => {
-    const rankingExtra = rankingMap.get(s.sku_codigo_omie);
-    const numCompras = rankingExtra?.num_compras_12m ?? null;
-    const mesesCompra = rankingExtra?.meses_com_compra ?? null;
-    const score = Number(s.score_final ?? 0);
-    return (
-      <Card key={s.id} className="flex flex-col">
-        <CardHeader className="pb-3">
-          <div className="flex items-start justify-between gap-2">
-            {s.categoria && s.categoria !== "fraco" ? (
-              <Badge variant="outline" className={cn("uppercase text-[10px] tracking-wide", categoriaBadgeClass(s.categoria))}>
-                {categoriaLabel(s.categoria)}
-              </Badge>
-            ) : <span />}
-            <Badge variant="outline" className={cn("text-[10px]", statusBadgeClass(s.status))}>
-              {statusLabel(s.status)}
-            </Badge>
-          </div>
-          <CardTitle className="text-base leading-snug mt-2 break-words">
-            {s.sku_descricao ?? "Sem descrição"}
-          </CardTitle>
-          <p className="text-xs font-mono text-muted-foreground">{s.sku_codigo_omie}</p>
-        </CardHeader>
-        <CardContent className="flex-1 flex flex-col gap-4 pt-0">
-          <div>
-            <div className="flex items-center justify-between text-xs mb-1.5">
-              <span className="font-medium">Score</span>
-              <span className="font-semibold">{score.toFixed(1)}</span>
-            </div>
-            <Progress value={score} className="h-2" />
-          </div>
-
-          {s.motivo && (
-            <p className="text-sm text-muted-foreground leading-relaxed">
-              {(s.motivo_detalhes && (s.motivo_detalhes as Record<string, unknown>).motivo_legivel as string | undefined) || s.motivo}
-            </p>
-          )}
-
-          <div className="grid grid-cols-2 gap-3 text-xs">
-            <div>
-              <p className="text-muted-foreground">Volume 12m</p>
-              <p className="font-semibold">{formatBRL(s.volume_financeiro_12m)}</p>
-            </div>
-            <div>
-              <p className="text-muted-foreground">Preço médio</p>
-              <p className="font-semibold">{formatBRL(s.preco_medio_unitario)}</p>
-            </div>
-            <div>
-              <p className="text-muted-foreground">Compras 12m</p>
-              <p className="font-semibold">
-                {numCompras ?? "—"}
-                {mesesCompra !== null && (
-                  <span className="text-muted-foreground font-normal"> em {mesesCompra} meses</span>
-                )}
-              </p>
-            </div>
-            <div>
-              <p className="text-muted-foreground">% meses com promo</p>
-              <p className="font-semibold">{formatPerc(s.perc_meses_com_promo)}</p>
-            </div>
-          </div>
-
-          <div className="flex items-center justify-between text-xs px-3 py-2 bg-muted/40 rounded-md border border-border">
-            <span className="text-muted-foreground">
-              Estoque: <span className="font-semibold text-foreground">{s.estoque_efetivo ?? 0}</span>
-            </span>
-            <span className="text-muted-foreground">
-              PP: <span className="font-semibold text-foreground">{s.ponto_pedido ?? "—"}</span>
-            </span>
-          </div>
-
-          <p className="text-xs text-muted-foreground">
-            Expira em {s.dias_ate_expirar ?? "—"} dia{(s.dias_ate_expirar ?? 0) === 1 ? "" : "s"}
-          </p>
-
-          <div className="flex flex-wrap items-center gap-2 pt-2 border-t border-border">
-            <Button
-              variant="outline"
-              size="sm"
-              disabled={s.status !== "nova"}
-              onClick={() => handleMarcarVisualizada(s)}
-            >
-              <Eye className="h-3.5 w-3.5 mr-1.5" />
-              Marcar visualizada
-            </Button>
-            <DropdownMenu>
-              <DropdownMenuTrigger asChild>
-                <Button variant="outline" size="sm">
-                  <MoreVertical className="h-3.5 w-3.5 mr-1.5" />
-                  Ações
-                </Button>
-              </DropdownMenuTrigger>
-              <DropdownMenuContent align="end">
-                <DropdownMenuItem onClick={() => handleIrAoRanking(s)}>Ir ao ranking</DropdownMenuItem>
-                <DropdownMenuItem onClick={() => handleMarcarEmAndamento(s)}>
-                  Marcar como em andamento
-                </DropdownMenuItem>
-                <DropdownMenuSeparator />
-                <DropdownMenuItem onClick={() => setIgnoreTarget(s)}>Ignorar</DropdownMenuItem>
-                <DropdownMenuItem onClick={() => setFecharSemAcordoTarget(s)}>
-                  Fechar sem acordo
-                </DropdownMenuItem>
-              </DropdownMenuContent>
-            </DropdownMenu>
-            <Button
-              size="sm"
-              className="ml-auto bg-status-success hover:bg-status-success/90 text-white"
-              onClick={() => openConvertDialog(s)}
-            >
-              <Handshake className="h-3.5 w-3.5 mr-1.5" />
-              Registrar desconto fechado
-            </Button>
-          </div>
-        </CardContent>
-      </Card>
-    );
-  };
-
   return (
     <div className="container mx-auto py-6 space-y-6 max-w-screen-2xl">
       {/* Breadcrumb + título */}
@@ -668,7 +518,19 @@ export default function AdminReposicaoNegociacaoParalela() {
           </Card>
         ) : (
           <div className="grid grid-cols-1 lg:grid-cols-2 gap-4">
-            {sugestoesFiltradas.map(renderSugestaoCard)}
+            {sugestoesFiltradas.map((s) => (
+              <SugestaoCard
+                key={s.id}
+                s={s}
+                rankingExtra={rankingMap.get(s.sku_codigo_omie)}
+                onMarcarVisualizada={handleMarcarVisualizada}
+                onIrAoRanking={handleIrAoRanking}
+                onMarcarEmAndamento={handleMarcarEmAndamento}
+                onIgnorar={(sug) => setIgnoreTarget(sug)}
+                onFecharSemAcordo={(sug) => setFecharSemAcordoTarget(sug)}
+                onConverter={openConvertDialog}
+              />
+            ))}
           </div>
         )}
       </section>
@@ -762,117 +624,15 @@ export default function AdminReposicaoNegociacaoParalela() {
         </div>
 
         {/* Tabela */}
-        <Card>
-          <CardContent className="p-0 overflow-x-auto">
-            <Table>
-              <TableHeader>
-                <TableRow>
-                  <TableHead className="w-12">#</TableHead>
-                  <TableHead className="w-[35%] min-w-[280px]">SKU</TableHead>
-                  <TableHead>Volume 12m</TableHead>
-                  <TableHead>Compras</TableHead>
-                  <TableHead>Preço médio</TableHead>
-                  <TableHead>% meses promo</TableHead>
-                  <TableHead className="w-[180px]">Score</TableHead>
-                  <TableHead>Categoria</TableHead>
-                  <TableHead>Ação</TableHead>
-                </TableRow>
-              </TableHeader>
-              <TableBody>
-                {loadingRanking ? (
-                  <TableRow>
-                    <TableCell colSpan={9} className="text-center py-10 text-muted-foreground">
-                      <Loader2 className="h-5 w-5 animate-spin inline mr-2" />
-                      Carregando ranking...
-                    </TableCell>
-                  </TableRow>
-                ) : rankingPagina_.length === 0 ? (
-                  <TableRow>
-                    <TableCell colSpan={9} className="text-center py-10 text-muted-foreground">
-                      Nenhum SKU encontrado.
-                    </TableCell>
-                  </TableRow>
-                ) : (
-                  rankingPagina_.map((r, idx) => {
-                    const posicao = (paginaAtual - 1) * PAGE_SIZE + idx + 1;
-                    const temSugestao = skusComSugestao.has(r.sku_codigo_omie);
-                    const isHighlight = highlightSku === r.sku_codigo_omie;
-                    const score = Number(r.score_final ?? 0);
-                    return (
-                      <TableRow
-                        key={r.sku_codigo_omie}
-                        className={cn(isHighlight && "bg-primary/10 transition-colors")}
-                      >
-                        <TableCell className="font-mono text-xs text-muted-foreground">
-                          {posicao}
-                        </TableCell>
-                        <TableCell>
-                          <div className="space-y-0.5">
-                            <p className="text-xs font-mono text-muted-foreground">
-                              {r.sku_codigo_omie}
-                            </p>
-                            <p className="text-sm whitespace-normal break-words leading-snug">
-                              {r.sku_descricao ?? "—"}
-                            </p>
-                          </div>
-                        </TableCell>
-                        <TableCell className="font-medium">
-                          {formatBRL(r.volume_financeiro_12m)}
-                        </TableCell>
-                        <TableCell>
-                          <div className="flex flex-col gap-0.5">
-                            <span className="font-medium">{r.num_compras_12m ?? "—"}</span>
-                            {r.meses_com_compra !== null && (
-                              <Badge variant="outline" className="text-[10px] w-fit">
-                                em {r.meses_com_compra} meses
-                              </Badge>
-                            )}
-                          </div>
-                        </TableCell>
-                        <TableCell>{formatBRL(r.preco_medio_unitario)}</TableCell>
-                        <TableCell>
-                          <Badge variant="outline" className={percPromoBadgeClass(r.perc_meses_com_promo)}>
-                            {formatPerc(r.perc_meses_com_promo)}
-                          </Badge>
-                        </TableCell>
-                        <TableCell>
-                          <div className="flex items-center gap-2">
-                            <span className="font-semibold w-10">{score.toFixed(1)}</span>
-                            <Progress value={score} className="h-1.5 flex-1" />
-                          </div>
-                        </TableCell>
-                        <TableCell>
-                          <Badge
-                            variant="outline"
-                            className={cn("uppercase text-[10px]", categoriaBadgeClass(r.categoria))}
-                          >
-                            {categoriaLabel(r.categoria)}
-                          </Badge>
-                        </TableCell>
-                        <TableCell>
-                          {temSugestao ? (
-                            <Badge variant="outline" className="text-[10px]">
-                              Já sugerido
-                            </Badge>
-                          ) : (
-                            <Button
-                              size="sm"
-                              variant="outline"
-                              onClick={() => handleCriarSugestaoDoRanking(r)}
-                            >
-                              <Sparkles className="h-3 w-3 mr-1" />
-                              Criar sugestão
-                            </Button>
-                          )}
-                        </TableCell>
-                      </TableRow>
-                    );
-                  })
-                )}
-              </TableBody>
-            </Table>
-          </CardContent>
-        </Card>
+        <RankingTable
+          rows={rankingPagina_}
+          loading={loadingRanking}
+          paginaAtual={paginaAtual}
+          pageSize={PAGE_SIZE}
+          skusComSugestao={skusComSugestao}
+          highlightSku={highlightSku}
+          onCriarSugestao={handleCriarSugestaoDoRanking}
+        />
 
         {/* Paginação */}
         <div className="flex items-center justify-between flex-wrap gap-3">
@@ -906,24 +666,14 @@ export default function AdminReposicaoNegociacaoParalela() {
       </section>
 
       {/* Dialog: ignorar */}
-      <AlertDialog open={!!ignoreTarget} onOpenChange={(o) => !o && setIgnoreTarget(null)}>
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>Ignorar sugestão?</AlertDialogTitle>
-            <AlertDialogDescription>
-              A sugestão será marcada como ignorada e removida da lista ativa. Você ainda poderá
-              gerar novas sugestões depois.
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <AlertDialogFooter>
-            <AlertDialogCancel>Cancelar</AlertDialogCancel>
-            <AlertDialogAction onClick={handleIgnorarConfirm}>Ignorar</AlertDialogAction>
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
+      <IgnorarDialog
+        open={!!ignoreTarget}
+        onOpenChange={(o) => !o && setIgnoreTarget(null)}
+        onConfirm={handleIgnorarConfirm}
+      />
 
       {/* Dialog: fechar sem acordo */}
-      <Dialog
+      <FecharSemAcordoDialog
         open={!!fecharSemAcordoTarget}
         onOpenChange={(o) => {
           if (!o) {
@@ -931,143 +681,22 @@ export default function AdminReposicaoNegociacaoParalela() {
             setFecharObs("");
           }
         }}
-      >
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Fechar sem acordo</DialogTitle>
-            <DialogDescription>
-              Registre o motivo do encerramento sem acordo. Útil para histórico futuro.
-            </DialogDescription>
-          </DialogHeader>
-          <div className="space-y-2">
-            <Label>Observação (opcional)</Label>
-            <Textarea
-              rows={4}
-              value={fecharObs}
-              onChange={(e) => setFecharObs(e.target.value)}
-              placeholder="Ex: Sayerlack não aceitou contraproposta, alegou margem apertada."
-            />
-          </div>
-          <DialogFooter>
-            <Button variant="outline" onClick={() => setFecharSemAcordoTarget(null)}>
-              Cancelar
-            </Button>
-            <Button onClick={handleFecharSemAcordoConfirm}>Confirmar</Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+        obs={fecharObs}
+        onObsChange={setFecharObs}
+        onCancel={() => setFecharSemAcordoTarget(null)}
+        onConfirm={handleFecharSemAcordoConfirm}
+      />
 
       {/* Dialog: registrar desconto fechado (converter) */}
-      <Dialog open={!!convertTarget} onOpenChange={(o) => !o && setConvertTarget(null)}>
-        <DialogContent className="max-w-2xl">
-          <DialogHeader>
-            <DialogTitle>Registrar desconto fechado</DialogTitle>
-            <DialogDescription>
-              Converte a sugestão em uma campanha flat condicional vinculada ao SKU{" "}
-              <span className="font-mono">{convertTarget?.sku_codigo_omie}</span>.
-            </DialogDescription>
-          </DialogHeader>
-          <div className="grid grid-cols-2 gap-4">
-            <div className="space-y-2">
-              <Label>Desconto percentual (%)</Label>
-              <Input
-                type="number"
-                min={1}
-                max={50}
-                step={0.5}
-                value={convertForm.desconto_perc}
-                onChange={(e) =>
-                  setConvertForm((f) => ({ ...f, desconto_perc: parseFloat(e.target.value) || 0 }))
-                }
-              />
-            </div>
-            <div className="space-y-2">
-              <Label>Volume mínimo</Label>
-              <Input
-                type="number"
-                min={0}
-                step={1}
-                value={convertForm.volume_minimo}
-                onChange={(e) =>
-                  setConvertForm((f) => ({ ...f, volume_minimo: parseFloat(e.target.value) || 0 }))
-                }
-              />
-            </div>
-            <div className="space-y-2">
-              <Label>Unidade do volume</Label>
-              <Select
-                value={convertForm.volume_unidade}
-                onValueChange={(v) =>
-                  setConvertForm((f) => ({ ...f, volume_unidade: v as typeof f.volume_unidade }))
-                }
-              >
-                <SelectTrigger>
-                  <SelectValue />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="reais">Reais (R$)</SelectItem>
-                  <SelectItem value="unidades">Unidades</SelectItem>
-                  <SelectItem value="kg">Quilos (kg)</SelectItem>
-                  <SelectItem value="litros">Litros</SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-            <div className="space-y-2">
-              <Label>Data fim</Label>
-              <Input
-                type="date"
-                value={convertForm.data_fim}
-                onChange={(e) => setConvertForm((f) => ({ ...f, data_fim: e.target.value }))}
-              />
-            </div>
-            <div className="space-y-2">
-              <Label>Responsável</Label>
-              <Input
-                value={convertForm.responsavel}
-                onChange={(e) => setConvertForm((f) => ({ ...f, responsavel: e.target.value }))}
-                placeholder="Nome do vendedor / contato"
-              />
-            </div>
-            <div className="space-y-2">
-              <Label>Canal</Label>
-              <Select
-                value={convertForm.canal}
-                onValueChange={(v) =>
-                  setConvertForm((f) => ({ ...f, canal: v as typeof f.canal }))
-                }
-              >
-                <SelectTrigger>
-                  <SelectValue />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="ligacao">Ligação</SelectItem>
-                  <SelectItem value="whatsapp">WhatsApp</SelectItem>
-                  <SelectItem value="email">E-mail</SelectItem>
-                  <SelectItem value="visita_presencial">Visita presencial</SelectItem>
-                  <SelectItem value="outro">Outro</SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-            <div className="col-span-2 space-y-2">
-              <Label>Observações (opcional)</Label>
-              <Textarea
-                rows={3}
-                value={convertForm.observacoes}
-                onChange={(e) => setConvertForm((f) => ({ ...f, observacoes: e.target.value }))}
-              />
-            </div>
-          </div>
-          <DialogFooter>
-            <Button variant="outline" onClick={() => setConvertTarget(null)}>
-              Cancelar
-            </Button>
-            <Button onClick={handleConverterConfirm} disabled={convertSubmitting}>
-              {convertSubmitting && <Loader2 className="h-4 w-4 mr-2 animate-spin" />}
-              Converter em campanha
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+      <ConverterDialog
+        target={convertTarget}
+        form={convertForm}
+        setForm={setConvertForm}
+        submitting={convertSubmitting}
+        onOpenChange={(o) => !o && setConvertTarget(null)}
+        onCancel={() => setConvertTarget(null)}
+        onConfirm={handleConverterConfirm}
+      />
     </div>
   );
 }
