+++ tsconfig.strict.json
@@ -372,6 +372,9 @@
     "src/pages/FinanceiroSync.tsx",
     "src/components/reposicao/negociacaoParalela/types.ts",
     "src/components/reposicao/negociacaoParalela/helpers.ts",
+    "src/components/reposicao/negociacaoParalela/SugestaoCard.tsx",
+    "src/components/reposicao/negociacaoParalela/RankingTable.tsx",
+    "src/components/reposicao/negociacaoParalela/dialogs.tsx",
     "src/pages/AdminMonthlyReports.tsx",
     "src/pages/NfeReceipt.tsx",
     "src/pages/ResetPassword.tsx",
